+++ packages/analytics-js-plugins/src/storageEncryption/index.ts
@@ -3,7 +3,7 @@ import { ApplicationState } from '../types/common';
 import { ExtensionPlugin } from '../types/plugins';
 import { fromBase64, toBase64 } from '../utilities/common';
 import { ENCRYPTION_PREFIX_V1, ENCRYPTION_PREFIX_V3 } from './constants';
-import { legacyDecrypt } from './utils';
+import { legacyDecrypt } from './legacyDecryptV1';
 
 const pluginName = 'StorageEncryption';
 
@@ -18,7 +18,7 @@ const StorageEncryption = (): ExtensionPlugin => ({
     },
     decrypt(value: string): string {
       if (value.startsWith(ENCRYPTION_PREFIX_V3)) {
-        return fromBase64(value.substring(ENCRYPTION_PREFIX_V3.length)) ?? value;
+        return fromBase64(value.substring(ENCRYPTION_PREFIX_V3.length));
       }
 
       // To remove this once all the persistent data is migrated to new encryption.
