+++ packages/analytics-js-plugins/src/storageEncryption/legacyDecryptV1.ts
