+++ packages/analytics-js/__mocks__/remotePlugins/StorageEncryption.ts
